move transformEventPayload
