+++ npm/src/directory-sync/utils.ts
@@ -1,6 +1,5 @@
-import type { Directory, DirectorySyncEvent, DirectorySyncEventType, Group, User } from '../typings';
+import type { User } from '../typings';
 import { DirectorySyncProviders, UserPatchOperation, GroupPatchOperation } from '../typings';
-import { transformUser, transformGroup, transformUserGroup } from './transform';
 import lodash from 'lodash';
 
 const parseGroupOperation = (operation: GroupPatchOperation) => {
@@ -51,38 +50,6 @@ const getDirectorySyncProviders = (): { [K: string]: string } => {
     acc[key] = value;
     return acc;
   }, {});
-};
-
-const transformEventPayload = (
-  event: DirectorySyncEventType,
-  payload: { directory: Directory; group?: Group | null; user?: User | null }
-): DirectorySyncEvent => {
-  const { directory, group, user } = payload;
-  const { tenant, product, id: directory_id } = directory;
-
-  const eventPayload = {
-    event,
-    tenant,
-    product,
-    directory_id,
-  } as DirectorySyncEvent;
-
-  // User events
-  if (['user.created', 'user.updated', 'user.deleted'].includes(event) && user) {
-    eventPayload['data'] = transformUser(user);
-  }
-
-  // Group events
-  if (['group.created', 'group.updated', 'group.deleted'].includes(event) && group) {
-    eventPayload['data'] = transformGroup(group);
-  }
-
-  // Group membership events
-  if (['group.user_added', 'group.user_removed'].includes(event) && user && group) {
-    eventPayload['data'] = transformUserGroup(user, group);
-  }
-
-  return eventPayload;
 };
 
 // Parse the PATCH request body and return the user attributes (both standard and custom)
@@ -162,7 +129,6 @@ const updateRawUserAttributes = (raw, attributes) => {
 export {
   parseGroupOperation,
   getDirectorySyncProviders,
-  transformEventPayload,
   parseUserPatchRequest,
   extractStandardUserAttributes,
   updateRawUserAttributes,
